--- OSBindings/Mac/Clock Signal/Documents/ElectronDocument.swift
@@ -1,30 +1,0 @@
-//
-//  ElectronDocument.swift
-//  Clock Signal
-//
-//  Created by Thomas Harte on 03/01/2016.
-//  Copyright © 2016 Thomas Harte. All rights reserved.
-//
-
-import Foundation
-import AudioToolbox
-
-class ElectronDocument: MachineDocument {
-
-/*	override func aspectRatio() -> NSSize {
-		return NSSize(width: 11.0, height: 10.0)
-	}
-
-	fileprivate let displayTypeUserDefaultsKey = "electron.displayType"
-	override func establishStoredOptions() {
-		super.establishStoredOptions()
-		let standardUserDefaults = UserDefaults.standard
-		standardUserDefaults.register(defaults: [
-			displayTypeUserDefaultsKey: 0,
-		])
-
-		let displayType = standardUserDefaults.integer(forKey: self.displayTypeUserDefaultsKey)
-		electron.useTelevisionOutput = (displayType == 1)
-		self.displayTypeButton?.selectItem(at: displayType)
-	}*/
-}
